feat: Enhance split volume display mode with improved divider visibility
- Increase divider line thickness and opacity for better visibility in split volume mode
- Improve dividing line from 2.0f width with 180 alpha to 2.5f width with 200 alpha
- Fix zoom sensitivity variable type from double to float for ImGui compatibility
- Add comprehensive tests for split volume functionality
- Ensure proper variable scoping for effective zoom factor in debug display
- Format code with clang-format following Google style

+++ dependencies/BTQ_Render_Engine/include/components/footprint_panel.hpp
@@ -1,39 +1,50 @@
 #pragma once
 
-#include "MarketMicrostructureRenderer.h"
-#include "panel_base.hpp"
-#include "data/unified_data_pipeline.hpp"  // For DataType enum
-#include "data/VolumeDataTypes.h"          // For VolumeAnalysisType and VolumeDataType enums
+#include <atomic>
 #include <memory>
 #include <vector>
-#include <atomic>
+
+#include "MarketMicrostructureRenderer.h"
+#include "data/VolumeDataTypes.h"          // For VolumeAnalysisType and VolumeDataType enums
+#include "data/unified_data_pipeline.hpp"  // For DataType enum
+#include "panel_base.hpp"
 
 namespace BTQuant {
 
 // Footprint Cell Structure for Exocharts-style visualization
 struct FootprintCell {
-  double x;             // Time position (X-axis)
-  double y;             // Price position (Y-axis)
-  double width;         // Cell width (time duration)
-  double height;        // Cell height (price range)
-  double bid_volume;    // Total bid volume
-  double ask_volume;    // Total ask volume
-  double delta;         // Delta (bid_volume - ask_volume)
-  uint32_t trade_count; // Number of trades
-  double vwap;          // Volume-weighted average price
-  uint32_t buy_trade_count;      // Number of buy trades
-  uint32_t sell_trade_count;     // Number of sell trades
-  double max_single_trade_volume; // Maximum single trade volume
-  uint64_t start_time_ns;        // Start timestamp in nanoseconds
-  uint64_t end_time_ns;          // End timestamp in nanoseconds
+  double x;                        // Time position (X-axis)
+  double y;                        // Price position (Y-axis)
+  double width;                    // Cell width (time duration)
+  double height;                   // Cell height (price range)
+  double bid_volume;               // Total bid volume
+  double ask_volume;               // Total ask volume
+  double delta;                    // Delta (bid_volume - ask_volume)
+  uint32_t trade_count;            // Number of trades
+  double vwap;                     // Volume-weighted average price
+  uint32_t buy_trade_count;        // Number of buy trades
+  uint32_t sell_trade_count;       // Number of sell trades
+  double max_single_trade_volume;  // Maximum single trade volume
+  uint64_t start_time_ns;          // Start timestamp in nanoseconds
+  uint64_t end_time_ns;            // End timestamp in nanoseconds
 
   // Constructor
-  FootprintCell(double x_pos, double y_pos, double w, double h, double bid_vol,
-                double ask_vol, uint32_t count, double vwap_price)
-      : x(x_pos), y(y_pos), width(w), height(h), bid_volume(bid_vol),
-        ask_volume(ask_vol), delta(bid_vol - ask_vol), trade_count(count),
-        vwap(vwap_price), buy_trade_count(0), sell_trade_count(0),
-        max_single_trade_volume(0.0), start_time_ns(0), end_time_ns(0) {}
+  FootprintCell(double x_pos, double y_pos, double w, double h, double bid_vol, double ask_vol,
+                uint32_t count, double vwap_price)
+      : x(x_pos),
+        y(y_pos),
+        width(w),
+        height(h),
+        bid_volume(bid_vol),
+        ask_volume(ask_vol),
+        delta(bid_vol - ask_vol),
+        trade_count(count),
+        vwap(vwap_price),
+        buy_trade_count(0),
+        sell_trade_count(0),
+        max_single_trade_volume(0.0),
+        start_time_ns(0),
+        end_time_ns(0) {}
 };
 
 // Number Formatting Options
@@ -46,9 +57,8 @@ enum class NumberFormat {
 };
 
 class FootprintPanel : public PanelBase {
-public:
-  FootprintPanel(const PanelConfig &config,
-                 RenderEngine::MarketMicrostructureRenderer *renderer);
+ public:
+  FootprintPanel(const PanelConfig& config, RenderEngine::MarketMicrostructureRenderer* renderer);
 
   void update(float dt) override;
   void render() override;
@@ -56,8 +66,7 @@ public:
   uint32_t get_symbol_id() const { return symbol_id_; }
   void set_symbol_id(uint32_t id) {
     symbol_id_ = id;
-    if (renderer_)
-      renderer_->setSymbol(id);
+    if (renderer_) renderer_->setSymbol(id);
   }
 
   // Configuration
@@ -84,7 +93,9 @@ public:
   Data::VolumeDataType getVolumeDataType() const { return volume_data_type_; }
 
   // Time aggregation type selection
-  void setTimeAggregationType(Data::TimeAggregationType agg_type) { time_aggregation_type_ = agg_type; }
+  void setTimeAggregationType(Data::TimeAggregationType agg_type) {
+    time_aggregation_type_ = agg_type;
+  }
   Data::TimeAggregationType getTimeAggregationType() const { return time_aggregation_type_; }
 
   // Volume-based and Tick-based aggregation parameters
@@ -94,7 +105,9 @@ public:
   int getTickBasedNTicks() const { return tick_based_n_ticks_; }
 
   // Price aggregation type selection
-  void setPriceAggregationType(Data::PriceAggregationType agg_type) { price_aggregation_type_ = agg_type; }
+  void setPriceAggregationType(Data::PriceAggregationType agg_type) {
+    price_aggregation_type_ = agg_type;
+  }
   Data::PriceAggregationType getPriceAggregationType() const { return price_aggregation_type_; }
 
   // Custom price aggregation value
@@ -113,8 +126,8 @@ public:
   void setZoomSensitivity(double sensitivity) { zoom_sensitivity_ = sensitivity; }
   double getZoomSensitivity() const { return zoom_sensitivity_; }
 
-private:
-  RenderEngine::MarketMicrostructureRenderer *renderer_;
+ private:
+  RenderEngine::MarketMicrostructureRenderer* renderer_;
   uint32_t symbol_id_ = 0;
 
   // Data Type Selection
@@ -124,14 +137,14 @@ private:
   Data::VolumeDataType volume_data_type_ = Data::VolumeDataType::Delta;
 
   // Grid Configuration (Exocharts-style: 60 columns × 100 rows)
-  int grid_cols_ = 60;  // Number of time columns (minutes)
-  int grid_rows_ = 100; // Number of price rows (ticks)
+  int grid_cols_ = 60;   // Number of time columns (minutes)
+  int grid_rows_ = 100;  // Number of price rows (ticks)
 
   // Visualization Options
   bool show_volume_labels_ = true;
   bool show_delta_indicator_ = true;
-  bool enable_volume_filter_ = false; // Whether volume filtering is enabled
-  float delta_threshold_ = 0.0f; // Threshold for delta coloring
+  bool enable_volume_filter_ = false;  // Whether volume filtering is enabled
+  float delta_threshold_ = 0.0f;       // Threshold for delta coloring
 
   // Number Formatting Options
   NumberFormat number_format_ = NumberFormat::ThousandsK;  // Default to K suffix
@@ -151,12 +164,12 @@ private:
   double custom_price_aggregation_value_ = 0.1;
 
   // Volume threshold for filtering cells
-  double volume_threshold_ = 0.0; // Default: no filtering
+  double volume_threshold_ = 0.0;  // Default: no filtering
   static constexpr double min_volume_threshold_ = 0.0;
   static constexpr double max_volume_threshold_ = 100000.0;
 
   // Zoom sensitivity for cell size adjustment
-  double zoom_sensitivity_ = 1.0; // Default: normal sensitivity
+  float zoom_sensitivity_ = 1.0f;  // Default: normal sensitivity
 
   // Data dirty flag for immediate rendering updates
   std::atomic<bool> data_dirty_{true};
@@ -165,27 +178,29 @@ private:
   std::vector<FootprintCell> cells_;
 
   // Rendering Helpers
-  ImU32 getCellColor(const FootprintCell &cell, double max_volume = 10000.0) const;
-  ImU32 getDeltaColor(const FootprintCell &cell, float alpha) const;
-  ImU32 getBuySellColor(const FootprintCell &cell, double max_volume, float alpha) const;
-  ImU32 getVolumeIntensityColor(const FootprintCell &cell, double max_volume, float alpha) const;
-  std::string getCellLabel(const FootprintCell &cell) const;
-  void renderCell(const FootprintCell &cell, ImDrawList *draw_list, double max_volume = 10000.0);
-  void renderCell(const FootprintCell &cell, ImDrawList *draw_list, double max_volume,
-                 const std::vector<FootprintCell> &diagonal_imbalances,
-                 const std::vector<FootprintCell> &stacked_imbalances,
-                 double zoom_factor = 1.0);
-  void renderFilteredCell(const FootprintCell &cell, ImDrawList *draw_list, double max_volume = 10000.0, double zoom_factor = 1.0);
-  std::string getCellTooltip(const FootprintCell &cell) const;
+  ImU32 getCellColor(const FootprintCell& cell, double max_volume = 10000.0) const;
+  ImU32 getDeltaColor(const FootprintCell& cell, float alpha) const;
+  ImU32 getBuySellColor(const FootprintCell& cell, double max_volume, float alpha) const;
+  ImU32 getVolumeIntensityColor(const FootprintCell& cell, double max_volume, float alpha) const;
+  std::string getCellLabel(const FootprintCell& cell) const;
+  void renderCell(const FootprintCell& cell, ImDrawList* draw_list, double max_volume = 10000.0);
+  void renderCell(const FootprintCell& cell, ImDrawList* draw_list, double max_volume,
+                  const std::vector<FootprintCell>& diagonal_imbalances,
+                  const std::vector<FootprintCell>& stacked_imbalances, double zoom_factor = 1.0);
+  void renderFilteredCell(const FootprintCell& cell, ImDrawList* draw_list,
+                          double max_volume = 10000.0, double zoom_factor = 1.0);
+  std::string getCellTooltip(const FootprintCell& cell) const;
 
   // Imbalance Detection
-  bool isDiagonalImbalance(const FootprintCell& cell, const std::vector<FootprintCell>& all_cells) const;
-  bool isStackedImbalance(const FootprintCell& cell, const std::vector<FootprintCell>& all_cells) const;
+  bool isDiagonalImbalance(const FootprintCell& cell,
+                           const std::vector<FootprintCell>& all_cells) const;
+  bool isStackedImbalance(const FootprintCell& cell,
+                          const std::vector<FootprintCell>& all_cells) const;
   void detectImbalances(const std::vector<FootprintCell>& cells,
-                       std::vector<FootprintCell>& diagonal_imbalances,
-                       std::vector<FootprintCell>& stacked_imbalances) const;
+                        std::vector<FootprintCell>& diagonal_imbalances,
+                        std::vector<FootprintCell>& stacked_imbalances) const;
 
-private:
+ private:
   // Number formatting helper
   static std::string formatNumber(double value, NumberFormat format, int decimal_places);
 
@@ -196,4 +211,4 @@ private:
   ImFont* getOrCreateMonospaceFont() const;
 };
 
-} // namespace BTQuant
+}  // namespace BTQuant
